refactor: group files into log drivers API
