+++ libs/ngworker/lumberjack/http-driver/src/lib/configuration/http-driver-root.module.ts
@@ -3,7 +3,7 @@ import { Inject, NgModule, NgZone, Optional, SkipSelf } from '@angular/core';
 
 import { LogDriverConfig, logDriverConfigToken, logDriverToken } from '@ngworker/lumberjack';
 
-import { HttpDriver } from '../http.driver';
+import { HttpDriver } from '../log-drivers/http.driver';
 
 import { httpDriverConfigToken } from './http-driver-config.token';
 import { HttpDriverConfig } from './http-driver.config';
