+++ libs/ngworker/lumberjack/http-driver/src/lib/log-drivers/http.driver.ts
@@ -3,10 +3,10 @@ import { Inject, Injectable, NgZone } from '@angular/core';
 
 import { LogDriver, LumberjackLevel } from '@ngworker/lumberjack';
 
-import { httpDriverConfigToken } from './configuration/http-driver-config.token';
-import { HttpDriverConfig } from './configuration/http-driver.config';
-import { HttpLogEntry } from './http-log-entry';
-import { retryWithDelay } from './retry-with-delay.operator';
+import { httpDriverConfigToken } from '../configuration/http-driver-config.token';
+import { HttpDriverConfig } from '../configuration/http-driver.config';
+import { HttpLogEntry } from '../http-log-entry';
+import { retryWithDelay } from '../retry-with-delay.operator';
 
 @Injectable()
 export class HttpDriver implements LogDriver {
